capture the empty CPy device filesystem
The 'blank' CircuitPython filesystem, as defined by the results of executing
the storage.erase_filesystem() command.  It consists of magic dotfiles for
various operating systems.

boot_out.txt should not be under git repo control.  It is not part of the
blank filesystem, but rather written at boottime by the CircuitPython
implementation.